move non-react code back into a single package again
