+++ package.json
@@ -1,12 +1,10 @@
 {
   "name": "plonk-monorepo",
   "version": "0.0.0",
-  "private": "true",
+  "private": true,
   "workspaces": [
-    "packages/utils",
-    "packages/core",
-    "packages/hooks",
-    "packages/plonk"
+    "packages/plonk",
+    "packages/hooks"
   ],
   "repository": {
     "type": "git",
@@ -22,12 +20,18 @@
     "test": "npm run test -ws --if-present"
   },
   "devDependencies": {
+    "@testing-library/react": "^14.2.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.12.3",
     "@typescript-eslint/eslint-plugin": "^7.5.0",
     "@typescript-eslint/parser": "^7.5.0",
-    "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^27.9.0",
+    "eslint": "^8.57.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest": "^29.7.0",
     "prettier": "^3.2.5",
+    "ts-jest": "^29.1.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.4.3"
   }
