--- packages/core/.npmignore
@@ -1,3 +1,0 @@
-src
-jest.config.js
-tsconfig.json
