--- packages/core/README.md
@@ -1,3 +1,0 @@
-# @prtcl/plonk-core
-
-Core timer and utils package for [plonk](https://github.com/prtcl/plonk)
