--- packages/core/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": [
-    ".eslintrc.js",
-    "__tests__",
-    "dist",
-    "jest.config.js",
-    "node_modules",
-    "src/**/*.spec.ts",
-    "src/**/*.test.ts"
-  ]
-}
