+++ packages/hooks/package.json
@@ -1,12 +1,15 @@
 {
   "name": "@prtcl/plonk-hooks",
-  "version": "1.0.0",
+  "version": "1.0.1-beta.0",
   "description": "React hook wrappers for plonk",
   "author": "Cory O'Brien <cory@prtcl.cc>",
   "license": "MIT",
-  "main": "dist/index.js",
+  "type": "module",
   "exports": {
-    ".": "./dist/index.js"
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
   },
   "repository": {
     "type": "git",
@@ -18,21 +21,10 @@
     "test": "npx jest --env=jsdom"
   },
   "dependencies": {
-    "@prtcl/plonk-core": "^1.0.0-alpha.2",
+    "@prtcl/plonk": "1.0.1-beta.0",
     "client-only": "^0.0.1"
   },
   "peerDependencies": {
     "react": "17 - 18"
-  },
-  "devDependencies": {
-    "@testing-library/react": "^14.2.2",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.12.3",
-    "@types/react": "^18.2.74",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "react": "^18.2.0",
-    "ts-jest": "^29.1.2",
-    "typescript": "^5.4.3"
   }
 }
