+++ packages/hooks/src/index.ts
@@ -2,5 +2,5 @@
 
 import 'client-only';
 
-export { default as useFrames } from './useFrames';
+export { default as useFrames, type UseFramesOptions } from './useFrames';
 export { default as useMetro, type UseMetroOptions } from './useMetro';
