+++ packages/plonk/package.json
@@ -1,22 +1,14 @@
 {
-  "name": "plonk",
-  "version": "2.0.0",
+  "name": "@prtcl/plonk",
+  "version": "1.0.1-beta.0",
   "description": "Tiny library that provides timers, envelopes, and random generators",
   "author": "Cory O'Brien <cory@prtcl.cc>",
   "license": "MIT",
-  "main": "dist/index.js",
+  "type": "module",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
-    },
-    "./hooks": {
-      "types": "./dist/hooks/index.d.ts",
-      "default": "./dist/hooks/index.js"
-    },
-    "./utils": {
-      "types": "./dist/utils/index.d.ts",
-      "default": "./dist/utils/index.js"
     }
   },
   "repository": {
@@ -25,14 +17,7 @@
   },
   "scripts": {
     "build": "npx tsc",
-    "clean": "rm -rf dist | true"
-  },
-  "dependencies": {
-    "@prtcl/plonk-core": "^1.0.0",
-    "@prtcl/plonk-hooks": "^1.0.0",
-    "@prtcl/plonk-utils": "^1.0.0"
-  },
-  "devDependencies": {
-    "typescript": "^5.4.3"
+    "clean": "rm -rf dist | true",
+    "test": "npx jest --env=node && npx jest --env=jsdom"
   }
 }
