+++ packages/plonk/src/__tests__/Scale.test.ts
@@ -1,4 +1,4 @@
-import Scale from '../Scale';
+import Scale from '../math/Scale';
 
 describe('Scale', () => {
   it('scales an incoming value from a given range to an output range', () => {
