+++ packages/plonk/src/__tests__/expo.test.ts
@@ -1,4 +1,4 @@
-import expo from '../expo';
+import expo from '../utils/expo';
 
 type TestCase = {
   from: number;
