+++ packages/plonk/src/__tests__/ms.test.ts
@@ -1,4 +1,4 @@
-import ms, { TimeFormat } from '../ms';
+import ms, { TimeFormat } from '../utils/ms';
 
 type TestCase = {
   val: string | number | undefined | null;
