+++ packages/plonk/src/index.ts
@@ -1,1 +1,3 @@
-export * from '@prtcl/plonk-core';
+export * from './math';
+export * from './timers';
+export * from './utils';
