+++ packages/plonk/src/math/index.ts
@@ -1,12 +1,5 @@
 export { default as Drunk, type DrunkOptions, type DrunkState } from './Drunk';
 export { default as Env, type EnvOptions, type EnvState } from './Env';
-export { default as Frames, type FramesOptions } from './Frames';
-export {
-  default as Metro,
-  type MetroOptions,
-  type TimerState,
-  type TimerCallback,
-} from './Metro';
 export { default as Rand, type RandOptions, type RandState } from './Rand';
 export {
   default as Scale,
