+++ packages/plonk/src/timers/index.ts
@@ -1,0 +1,6 @@
+export { default as Frames, type FramesOptions } from './Frames';
+export {
+  default as Metro,
+  type MetroOptions,
+  type TimerCallback,
+} from './Metro';
