+++ packages/plonk/src/utils/index.ts
@@ -1,1 +1,6 @@
-export * from '@prtcl/plonk-utils';
+export { default as clamp } from './clamp';
+export { default as expo } from './expo';
+export { default as flip } from './flip';
+export { default as ms, type FPS, TimeFormat } from './ms';
+export { default as now } from './now';
+export * from './constants';
