--- packages/utils/README.md
@@ -1,3 +1,0 @@
-# @prtcl/plonk-utils
-
-Utils and helpers for [plonk](https://github.com/prtcl/plonk)
