--- packages/utils/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@prtcl/plonk-utils",
-  "version": "1.0.0",
-  "description": "Utils and helpers for plonk",
-  "author": "Cory O'Brien <cory@prtcl.cc>",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "exports": {
-    ".": "./dist/index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/prtcl/plonk.git"
-  },
-  "scripts": {
-    "build": "npx tsc",
-    "clean": "rm -rf dist | true",
-    "test": "npx jest --env=node && npx jest --env=jsdom"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.12.3",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "ts-jest": "^29.1.2",
-    "typescript": "^5.4.3"
-  }
-}
